+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Requires `genbase>=0.1.17`
 - Renamed `pyproject.toml` to `.portray` to avoid build errors
+- Made `fastcountvectorizer` optional
 
 ### Fixed
 - Bugfix when installing package, by moving `__version___` to `/_version.py`